+++ modulelibrary/src/main/java/cn/net/aicare/modulelibrary/module/BodyFatScale/BodyFatBleUtilsData.java
@@ -72,9 +72,7 @@ public class BodyFatBleUtilsData extends BaseBleDeviceData {
         //通用指令的回调
         mBleDevice.setOnBleSettingListener(new OnBleSettingListener() {
             @Override
-            public void OnSettingReturn(byte cmdType, byte cmdData) {
-
-
+            public void OnSettingReturn(int cmdType, int cmdData) {
                 if (mBleBodyFatCallback != null)
                     mBleBodyFatCallback.setTimeCallback(cmdType, cmdData);
             }
